Move check ad and check cmp functions to utils
